+++ web/wasm24/wasm24.d.ts
@@ -2,13 +2,13 @@
 /* eslint-disable */
 /**
 * @param {number} target
-* @param {Uint16Array} list
+* @param {Int32Array} list
 * @returns {string | undefined}
 */
-export function solve(target: number, list: Uint16Array): string | undefined;
+export function solve(target: number, list: Int32Array): string | undefined;
 /**
 * @param {number} target
-* @param {Uint16Array} list
+* @param {Int32Array} list
 * @returns {(string)[]}
 */
-export function solve_all(target: number, list: Uint16Array): (string)[];
+export function solve_all(target: number, list: Int32Array): (string)[];
